requirements.txt
asgiref==3.8.1
Django==5.2
sqlparse==0.5.3
typing_extensions==4.13.1
djangorestframework>=3.12.0,<4.0.0
psycopg2-binary>=2.9.0,<3.0.0
dj-database-url>=0.5.0,<1.0.0
